requirements.txt
asgiref==3.7.2
Django==5.0.2
django-mathfilters==1.0.0
grpcio==1.59.3
idna==3.6
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pillow==10.2.0
pivottablejs==0.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
Pygments==2.17.2
pyparsing==3.1.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tensorboard-data-server==0.7.2
tornado==6.4
traitlets==5.14.1
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
mysql-connector-python==8.4.0